# Create less garbage when parsing metrics