I noticed that the profile for #9253 had 94% of all allocations in `scrapeLoop.append()`:
```
140179748.80MB 94.40% 94.40% 141080447.83MB 95.01%  github.com/prometheus/prometheus/scrape.(*scrapeLoop).append
```

so added benchmarks and found the cause - Exemplar variable escapes to heap due to being passed through text-parser
interface.  (Note the cost was paid on every line even when exemplars were disabled)

In a different Prometheus I checked this line contributes only 6%, but there is another in the OpenMetrics parser which costs about the same, so I optimised that too.
After calling `parseLVals()` we always append the return value, so pass in what we want to append it to and save garbage.

Benchmark comparison:
```
name                  old time/op    new time/op    delta
ScrapeLoopAppend-8      54.6µs ± 1%    49.3µs ± 2%   -9.60%  (p=0.008 n=5+5)
ScrapeLoopAppendOM-8    58.8µs ± 2%    46.0µs ± 3%  -21.75%  (p=0.008 n=5+5)

name                  old alloc/op   new alloc/op   delta
ScrapeLoopAppend-8      34.5kB ± 2%    21.3kB ± 8%  -38.40%  (p=0.008 n=5+5)
ScrapeLoopAppendOM-8    46.5kB ± 0%    20.3kB ±10%  -56.30%  (p=0.008 n=5+5)

name                  old allocs/op  new allocs/op  delta
ScrapeLoopAppend-8         409 ± 0%       109 ± 0%  -73.35%  (p=0.008 n=5+5)
ScrapeLoopAppendOM-8       709 ± 0%       111 ± 0%  -84.34%  (p=0.008 n=5+5)
```